Added template .env file.
This is a convenience for running an app on Heroku. To configure a
Heroku app you use environment variables rather than values in config
files. To replicate this locally, you can add values into a .env file in
your app root, and then when you run the app using `foreman start`
foreman makes sure the env variables are correctly set.

Alternatively you can manually set them either globally or in the
current terminal. This is a bit of a faff, so the .env and foreman
option is probably easiest.